[feature](selectdb-cloud) Fix the coredump when eliminating cache because encounting query limit (#1583)
The pr fix three problems.
1. When eliminating ghost, it will cause heap-use-after-free.
2. When adding a segment twice, it will coredump because of decheck.
3.  If the disposable path is same as the cache path , it will rm the cache in disk.

+++ be/src/service/doris_main.cpp
@@ -27,6 +27,7 @@
 #include <thread>
 #include <unordered_map>
 
+#include "cloud/io/cloud_file_cache_profile.h"
 #include "util/jni-util.h"
 
 #if defined(LEAK_SANITIZER)
@@ -405,8 +406,8 @@ int main(int argc, char** argv) {
         }
         for (auto& cache_path : cache_paths) {
             if (cache_path_set.find(cache_path.path) != cache_path_set.end()) {
-                LOG(FATAL) << fmt::format("cache path {} is duplicate", cache_path.path);
-                exit(-1);
+                LOG(WARNING) << fmt::format("cache path {} is duplicate", cache_path.path);
+                continue;
             }
             Status st = doris::io::FileCacheFactory::instance().create_file_cache(
                     cache_path.path, cache_path.init_settings());
@@ -415,6 +416,8 @@ int main(int argc, char** argv) {
                 exit(-1);
             }
         }
+        // use to init the 0 table profile for cloud mgr
+        doris::io::FileCacheProfile::instance();
         std::vector<std::string> rm_paths;
         olap_res = doris::parse_conf_rm_paths(doris::config::disposable_file_cache_path, rm_paths);
         if (!olap_res) {
@@ -422,8 +425,12 @@ int main(int argc, char** argv) {
                        << doris::config::disposable_file_cache_path;
             exit(-1);
         }
-        std::for_each(rm_paths.begin(), rm_paths.end(), [](const std::string& path) {
-            doris::io::global_local_filesystem()->delete_directory(path);
+        std::for_each(rm_paths.begin(), rm_paths.end(), [&cache_path_set](const std::string& path) {
+            if (cache_path_set.find(path) == cache_path_set.end()) {
+                doris::io::global_local_filesystem()->delete_directory(path);
+            } else {
+                LOG(WARNING) << fmt::format("disposable cache path {} is duplicate.", path);
+            }
         });
     }
